Declarations of upload simple and multipart were separated
